+++ CHANGELOG.md
@@ -2,11 +2,13 @@
 
 FEATURES:
 
+* **New Resource**: `tencentcloud_sqlserver_publish_subscribe`
 * **New Resource**: `tencentcloud_vod_adaptive_dynamic_streaming_template`.
 * **New Resource**: `tencentcloud_vod_procedure_template`.
 * **New Resource**: `tencentcloud_vod_snapshot_by_time_offset_template`.
 * **New Resource**: `tencentcloud_vod_image_sprite_template`.
 * **New Resource**: `tencentcloud_vod_super_player_config`.
+* **New Data Source**: `tencentcloud_sqlserver_publish_subscribes` 
 * **New Data Source**: `tencentcloud_vod_adaptive_dynamic_streaming_templates`.
 * **New Data Source**: `tencentcloud_vod_image_sprite_templates`.
 * **New Data Source**: `tencentcloud_vod_procedure_templates`.